refactor: Standardize module naming `src/xblock_<foo>/block.py`
I made the app changes, and Claude fixed the mocks.
