+++ src/xblock_lti/tests/test_lti_unit.py
@@ -17,7 +17,7 @@ from webob.request import Request
 from xblock.field_data import DictFieldData
 from xblock.fields import ScopeIds
 
-from xblock_lti.lti import LTIBlock
+from xblock_lti.block import LTIBlock
 from xblock_lti.lti_2_util import LTIError
 
 from .helpers import StubUserService, Timedelta, get_test_system
@@ -135,7 +135,7 @@ class LTIBlockTest(TestCase):
         }
 
     @patch(
-        "xblock_lti.lti.LTIBlock.get_client_key_secret",
+        "xblock_lti.block.LTIBlock.get_client_key_secret",
         return_value=("test_client_key", "test_client_secret"),
     )
     def test_authorization_header_not_present(self, _get_key_secret):
@@ -159,7 +159,7 @@ class LTIBlockTest(TestCase):
         self.assertDictEqual(expected_response, real_response)
 
     @patch(
-        "xblock_lti.lti.LTIBlock.get_client_key_secret",
+        "xblock_lti.block.LTIBlock.get_client_key_secret",
         return_value=("test_client_key", "test_client_secret"),
     )
     def test_authorization_header_empty(self, _get_key_secret):
@@ -326,7 +326,7 @@ class LTIBlockTest(TestCase):
 
     def test_resource_link_id(self):
         with patch(
-            "xblock_lti.lti.LTIBlock.usage_key", new_callable=PropertyMock
+            "xblock_lti.block.LTIBlock.usage_key", new_callable=PropertyMock
         ):
             self.xblock.usage_key.html_id = (
                 lambda: "i4x-2-3-lti-31de800015cf4afb973356dbe81496df"
@@ -398,10 +398,10 @@ class LTIBlockTest(TestCase):
             self.xblock.get_client_key_secret()
 
     @patch(
-        "xblock_lti.lti.signature.verify_hmac_sha1", Mock(return_value=True)
+        "xblock_lti.block.signature.verify_hmac_sha1", Mock(return_value=True)
     )
     @patch(
-        "xblock_lti.lti.LTIBlock.get_client_key_secret",
+        "xblock_lti.block.LTIBlock.get_client_key_secret",
         Mock(return_value=("test_client_key", "test_client_secret")),
     )
     def test_successful_verify_oauth_body_sign(self):
@@ -411,11 +411,11 @@ class LTIBlockTest(TestCase):
         self.xblock.verify_oauth_body_sign(self.get_signed_grade_mock_request())
 
     @patch(
-        "xblock_lti.lti.LTIBlock.get_outcome_service_url",
+        "xblock_lti.block.LTIBlock.get_outcome_service_url",
         Mock(return_value="https://testurl/"),
     )
     @patch(
-        "xblock_lti.lti.LTIBlock.get_client_key_secret",
+        "xblock_lti.block.LTIBlock.get_client_key_secret",
         Mock(return_value=("__consumer_key__", "__lti_secret__")),
     )
     def test_failed_verify_oauth_body_sign_proxy_mangle_url(self):
@@ -491,10 +491,10 @@ class LTIBlockTest(TestCase):
         assert self.defaults["action"] == action
 
     @patch(
-        "xblock_lti.lti.signature.verify_hmac_sha1", Mock(return_value=False)
+        "xblock_lti.block.signature.verify_hmac_sha1", Mock(return_value=False)
     )
     @patch(
-        "xblock_lti.lti.LTIBlock.get_client_key_secret",
+        "xblock_lti.block.LTIBlock.get_client_key_secret",
         Mock(return_value=("test_client_key", "test_client_secret")),
     )
     def test_failed_verify_oauth_body_sign(self):
