+++ src/xblock_pdf/__init__.py
@@ -1,3 +1,3 @@
 """Init for PDFBlock."""
 
-from .pdf import PDFBlock
+from .block import PDFBlock
