+++ src/xblock_problem/__init__.py
@@ -2,4 +2,4 @@
 Init for the ProblemBlock.
 """
 
-from .capa_block import ProblemBlock
+from .block import ProblemBlock
