+++ src/xblock_video/tests/__init__.py
@@ -4,6 +4,6 @@ Tests for the video xblock.
 
 from unittest.mock import Mock
 
-from ..video import VideoBlock
+from ..block import VideoBlock
 
 VideoBlock.add_aside = Mock()
