+++ src/xblock_word_cloud/__init__.py
@@ -1,4 +1,4 @@
 """Word cloud is ungraded xblock used by students to generate and view word cloud.
 """
 
-from .word_cloud import WordCloudBlock
+from .block import WordCloudBlock
